AI/requirements.txt
python-dotenv
openai>=1.0.0
langgraph
typing_extensions
weasyprint
jinja2